Deprecate old op-rbuilder
