+++ Cargo.toml
@@ -4,7 +4,6 @@ members = [
     "crates/op-rbuilder",
     "crates/op-rbuilder/payload_builder",
     "crates/op-rbuilder/node",
-    "crates/op-rbuilder2",
     "crates/reth-rbuilder",
     "crates/rbuilder/src/test_utils",
     "crates/rbuilder/src/telemetry/metrics_macros",
