+++ Makefile
@@ -38,13 +38,11 @@ lint: ## Run the linters
 	cargo fmt -- --check
 	cargo clippy --features "$(FEATURES)" -- -D warnings
 	cargo clippy -p op-rbuilder --features "$(FEATURES),optimism" -- -D warnings
-	cargo clippy -p op-rbuilder2 --features "$(FEATURES),optimism" -- -D warnings
 
 .PHONY: test
 test: ## Run the tests for rbuilder and op-rbuilder
 	cargo test --verbose --features "$(FEATURES)"
 	cargo test -p op-rbuilder --verbose --features "$(FEATURES),optimism"
-	cargo test -p op-rbuilder2 --verbose --features "$(FEATURES),optimism"
 
 .PHONY: lt
 lt: lint test ## Run "lint" and "test"
@@ -55,7 +53,6 @@ fmt: ## Format the code
 	cargo fix --allow-staged
 	cargo clippy --features "$(FEATURES)" --fix --allow-staged
 	cargo clippy -p op-rbuilder --features "$(FEATURES),optimism" --fix --allow-staged
-	cargo clippy -p op-rbuilder2 --features "$(FEATURES),optimism" --fix --allow-staged
 
 .PHONY: bench
 bench: ## Run benchmarks
