+++ crates/op-rbuilder/Cargo.toml
@@ -1,23 +1,63 @@
 [package]
-name = "op-rbuilder"
+name = "op-rbuilder2"
 version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-rbuilder = { path = "../rbuilder" }
-op-rbuilder-node-optimism = { path = "./node" }
-transaction-pool-bundle-ext = { path = "../transaction-pool-bundle-ext" }
-
 reth.workspace = true
 reth-optimism-node.workspace = true
 reth-optimism-cli.workspace = true
+reth-optimism-chainspec.workspace = true
+reth-optimism-payload-builder.workspace = true
+reth-optimism-evm.workspace = true
+reth-optimism-consensus.workspace = true
 reth-cli-util.workspace = true
+reth-payload-primitives.workspace = true
+reth-evm.workspace = true
+reth-chainspec.workspace = true
+reth-primitives.workspace = true
+reth-node-api.workspace = true
+reth-basic-payload-builder.workspace = true
+reth-payload-builder.workspace = true
+reth-node-ethereum.workspace = true
+reth-chain-state.workspace = true
+reth-execution-types.workspace = true
+reth-ethereum-payload-builder.workspace = true
+reth-provider.workspace = true
+reth-revm.workspace = true
+reth-trie.workspace = true
+reth-rpc-layer.workspace = true
+reth-payload-builder-primitives.workspace = true
+reth-payload-util.workspace = true
+reth-transaction-pool.workspace = true
+reth-optimism-forks.workspace = true
+alloy-primitives.workspace = true
+alloy-consensus.workspace = true
+alloy-eips.workspace = true
+alloy-rpc-types-beacon.workspace = true
+alloy-rpc-types-engine.workspace = true
+op-alloy-consensus.workspace = true
+op-alloy-provider.workspace = true
+op-alloy-rpc-types-engine.workspace = true
+alloy-transport-http.workspace = true
+alloy-rpc-types-eth.workspace = true
+alloy-rpc-client.workspace = true
+alloy-transport.workspace = true
+alloy-network.workspace = true
+
+revm.workspace = true
+
+tracing.workspace = true
+futures-util = "0.3.31"
+eyre.workspace = true
+alloy-provider.workspace = true
+tower = "0.4"
 
 tokio.workspace = true
-tracing.workspace = true
 jsonrpsee = { workspace = true }
 async-trait = { workspace = true }
 clap_builder = { workspace = true }
+clap.workspace = true
 
 [target.'cfg(unix)'.dependencies]
 tikv-jemallocator = { version = "0.6", optional = true }
@@ -47,16 +87,14 @@ min-debug-logs = ["tracing/release_max_level_debug"]
 min-trace-logs = ["tracing/release_max_level_trace"]
 
 optimism = [
-	"rbuilder/optimism",
 	"reth-optimism-node/optimism",
-	"op-rbuilder-node-optimism/optimism",
 	"reth-optimism-cli/optimism"
 ]
 
-redact-sensitive = [
-	"rbuilder/redact-sensitive"
-]
+[[bin]]
+name = "op-rbuilder2"
+path = "src/main.rs"
 
 [[bin]]
-name = "op-rbuilder"
-path = "src/main.rs"
+name = "tester"
+path = "src/tester/tester.rs"
