--- crates/op-rbuilder/node/Cargo.toml
@@ -1,44 +1,0 @@
-[package]
-name = "op-rbuilder-node-optimism"
-edition = "2021"
-
-[dependencies]
-# workspace
-op-rbuilder-payload-builder = { path = "../payload_builder" }
-transaction-pool-bundle-ext = { path = "../../transaction-pool-bundle-ext" }
-rbuilder-bundle-pool-operations = { path = "../../transaction-pool-bundle-ext/bundle_pool_ops/rbuilder" }
-
-# alloy
-alloy-consensus.workspace = true
-
-# reth
-reth-payload-builder.workspace = true
-reth-primitives.workspace = true
-reth-basic-payload-builder.workspace = true
-reth-node-builder.workspace = true
-reth-node-api.workspace = true
-reth-trie-db.workspace = true
-reth-tracing.workspace = true
-reth-provider.workspace = true
-reth-transaction-pool.workspace = true
-reth-evm.workspace = true
-reth-optimism-evm.workspace = true
-reth-optimism-chainspec.workspace = true
-reth-optimism-node = { workspace = true }
-
-# async
-tracing.workspace = true
-
-# misc
-clap.workspace = true
-eyre.workspace = true
-
-[features]
-optimism = [
-	"rbuilder-bundle-pool-operations/optimism",
-	"reth-optimism-evm/optimism",
-	"reth-optimism-node/optimism",
-	"reth-primitives/optimism",
-	"reth-provider/optimism",
-	"op-rbuilder-payload-builder/optimism",
-]
