--- crates/op-rbuilder/payload_builder/Cargo.toml
@@ -1,50 +1,0 @@
-[package]
-name = "op-rbuilder-payload-builder"
-version = "0.1.0"
-edition = "2021"
-description = "A payload builder for op-rbuilder with bundle support."
-
-[dependencies]
-# workspace
-transaction-pool-bundle-ext = { path = "../../transaction-pool-bundle-ext" }
-
-# reth
-reth-chainspec.workspace = true
-reth-primitives.workspace = true
-reth-revm.workspace = true
-reth-provider.workspace = true
-reth-evm.workspace = true
-reth-optimism-evm.workspace = true
-reth-optimism-consensus.workspace = true
-reth-optimism-chainspec.workspace = true
-reth-optimism-forks.workspace = true
-reth-optimism-node.workspace = true
-reth-execution-types.workspace = true
-reth-payload-builder.workspace = true
-reth-payload-primitives.workspace = true
-reth-basic-payload-builder.workspace = true
-reth-trie.workspace = true
-reth-chain-state.workspace = true
-reth-optimism-payload-builder.workspace = true
-
-# ethereum
-revm.workspace = true
-alloy-consensus.workspace = true
-op-alloy-consensus.workspace = true
-alloy-eips.workspace = true
-alloy-rpc-types-engine.workspace = true
-alloy-rpc-types-beacon.workspace = true
-
-# misc
-tracing.workspace = true
-
-[features]
-optimism = [
-	"reth-execution-types/optimism",
-	"reth-optimism-evm/optimism",
-	"reth-optimism-node/optimism",
-	"reth-primitives/optimism",
-	"reth-provider/optimism",
-	"revm/optimism",
-	"reth-optimism-consensus/optimism",
-]
