+++ crates/op-rbuilder/src/generator.rs
