+++ crates/op-rbuilder/src/main.rs
@@ -1,84 +1,89 @@
-//! The main entry point for `op-rbuilder`.
-//!
-//! `op-rbuilder` is an OP Stack EL client with block building capabilities, powered by in-process
-//! rbuilder.
-//!
-//! The primary difference between `op-rbuilder` and `op-reth` is the `PayloadBuilder` derives
-//! transactions exclusively from rbuilder, rather than directly from its transaction pool.
-//!
-//! ## Usage
-//!
-//! It has a new mandatory cli arg `--rbuilder.config` which must point to an rbuilder config file.
-//!
-//! ## Demo
-//!
-//! Instructions to demo `op-rbuilder` building blocks for an OP L2, and send txns to it with `mev-flood`:
-//!
-//! 1. Clone [flashbots/optimism](https://github.com/flashbots/optimism) and checkout the
-//!    `op-rbuilder` branch.
-//! 2. `rm` any existing `reth` chain db
-//! 3. Run a clean OP stack: `make devnet-clean && make devnet-down && make devnet-up`
-//! 4. Run `op-rbuilder` on port 8547: `cargo run --bin op-rbuilder --features "optimism,jemalloc" -- node
-//!    --chain ../optimism/.devnet/genesis-l2.json --http --http.port 8547 --authrpc.jwtsecret
-//!    ../optimism/ops-bedrock/test-jwt-secret.txt --rbuilder.config config-optimism-local.toml`
-//! 5. Init `mev-flood`: `docker run mevflood init -r http://host.docker.internal:8547 -s local.json`
-//! 6. Run `mev-flood`: `docker run --init -v ${PWD}:/app/cli/deployments mevflood spam -p 3 -t 5 -r http://host.docker.internal:8547 -l local.json`
-//!
-//! Example starting clean OP Stack in one-line: `rm -rf /Users/liamaharon/Library/Application\ Support/reth && cd ../optimism && make devnet-clean && make devnet-down && make devnet-up && cd ../rbuilder && cargo run --bin op-rbuilder --features "optimism,jemalloc" -- node --chain ../optimism/.devnet/genesis-l2.json --http --http.port 8547 --authrpc.jwtsecret ../optimism/ops-bedrock/test-jwt-secret.txt --rbuilder.config config-optimism-local.toml`
-
-#![cfg_attr(all(not(test), feature = "optimism"), warn(unused_crate_dependencies))]
-// The `optimism` feature must be enabled to use this crate.
-#![cfg(feature = "optimism")]
-
-mod eth_bundle_api;
-
-use crate::eth_bundle_api::EthCallBundleMinimalApiServer;
-use clap_builder::Parser;
-use eth_bundle_api::EthBundleMinimalApi;
-use op_rbuilder_node_optimism::{args::OpRbuilderArgs, OpRbuilderNode};
+use clap::Parser;
+use generator::EmptyBlockPayloadJobGenerator;
+use payload_builder::OpPayloadBuilder as FBPayloadBuilder;
+use payload_builder_vanilla::VanillaOpPayloadBuilder;
+use reth::{
+    builder::{components::PayloadServiceBuilder, node::FullNodeTypes, BuilderContext},
+    payload::PayloadBuilderHandle,
+    providers::CanonStateSubscriptions,
+    transaction_pool::TransactionPool,
+};
 use reth::{
     builder::{engine_tree_config::TreeConfig, EngineNodeLauncher},
     providers::providers::BlockchainProvider2,
 };
+use reth_basic_payload_builder::BasicPayloadJobGeneratorConfig;
+use reth_node_api::NodeTypesWithEngine;
+use reth_optimism_chainspec::OpChainSpec;
 use reth_optimism_cli::{chainspec::OpChainSpecParser, Cli};
-use reth_optimism_node::node::OpAddOns;
-use tracing as _;
+use reth_optimism_evm::OpEvmConfig;
+use reth_optimism_node::OpEngineTypes;
+use reth_optimism_node::{args::RollupArgs, node::OpAddOns, OpNode};
+use reth_payload_builder::PayloadBuilderService;
 
-// jemalloc provides better performance
-#[cfg(all(feature = "jemalloc", unix))]
-#[global_allocator]
-static ALLOC: tikv_jemallocator::Jemalloc = tikv_jemallocator::Jemalloc;
+pub mod generator;
+pub mod payload_builder;
+mod payload_builder_vanilla;
+
+#[derive(Debug, Clone, Copy, Default)]
+#[non_exhaustive]
+pub struct CustomPayloadBuilder;
+
+impl<Node, Pool> PayloadServiceBuilder<Node, Pool> for CustomPayloadBuilder
+where
+    Node:
+        FullNodeTypes<Types: NodeTypesWithEngine<Engine = OpEngineTypes, ChainSpec = OpChainSpec>>,
+    Pool: TransactionPool + Unpin + 'static,
+{
+    async fn spawn_payload_service(
+        self,
+        ctx: &BuilderContext<Node>,
+        pool: Pool,
+    ) -> eyre::Result<PayloadBuilderHandle<<Node::Types as NodeTypesWithEngine>::Engine>> {
+        tracing::info!("Spawning a custom payload builder");
+        let _fb_builder = FBPayloadBuilder::new(OpEvmConfig::new(ctx.chain_spec()));
+        let vanilla_builder = VanillaOpPayloadBuilder::new(OpEvmConfig::new(ctx.chain_spec()));
+        let payload_job_config = BasicPayloadJobGeneratorConfig::default();
+
+        let payload_generator = EmptyBlockPayloadJobGenerator::with_builder(
+            ctx.provider().clone(),
+            pool,
+            ctx.task_executor().clone(),
+            payload_job_config,
+            // FBPayloadBuilder::new(OpEvmConfig::new(ctx.chain_spec())),
+            vanilla_builder,
+        );
+
+        let (payload_service, payload_builder) =
+            PayloadBuilderService::new(payload_generator, ctx.provider().canonical_state_stream());
+
+        ctx.task_executor()
+            .spawn_critical("custom payload builder service", Box::pin(payload_service));
+
+        Ok(payload_builder)
+    }
+}
 
 fn main() {
-    reth_cli_util::sigsegv_handler::install();
-
-    if std::env::var_os("RUST_BACKTRACE").is_none() {
-        std::env::set_var("RUST_BACKTRACE", "1");
-    }
-
-    if let Err(err) =
-        Cli::<OpChainSpecParser, OpRbuilderArgs>::parse().run(|builder, op_rbuilder_args| async move {
-            if op_rbuilder_args.experimental {
+    Cli::<OpChainSpecParser, RollupArgs>::parse()
+        .run(|builder, rollup_args| async move {
+            if rollup_args.experimental {
                 tracing::warn!(target: "reth::cli", "Experimental engine is default now, and the --engine.experimental flag is deprecated. To enable the legacy functionality, use --engine.legacy.");
             }
-            let use_legacy_engine = op_rbuilder_args.legacy;
-            let sequencer_http_arg = op_rbuilder_args.sequencer_http.clone();
+            let use_legacy_engine = rollup_args.legacy;
+            let sequencer_http_arg = rollup_args.sequencer_http.clone();
+
             match use_legacy_engine {
                 false => {
                     let engine_tree_config = TreeConfig::default()
-                        .with_persistence_threshold(op_rbuilder_args.persistence_threshold)
-                        .with_memory_block_buffer_target(op_rbuilder_args.memory_block_buffer_target);
+                        .with_persistence_threshold(rollup_args.persistence_threshold)
+                        .with_memory_block_buffer_target(rollup_args.memory_block_buffer_target);
                     let handle = builder
-                        .with_types_and_provider::<OpRbuilderNode, BlockchainProvider2<_>>()
-                        .with_components(OpRbuilderNode::components(op_rbuilder_args))
+                        .with_types_and_provider::<OpNode, BlockchainProvider2<_>>()
+                        .with_components(
+                            OpNode::components(rollup_args).payload(CustomPayloadBuilder::default()),
+                        )
                         .with_add_ons(OpAddOns::new(sequencer_http_arg))
-                        .extend_rpc_modules(move |ctx| {
-                            // register eth bundle api
-                            let ext = EthBundleMinimalApi::new(ctx.registry.pool().clone());
-                            ctx.modules.merge_configured(ext.into_rpc())?;
-
-                            Ok(())
-                        })
                         .launch_with_fn(|builder| {
                             let launcher = EngineNodeLauncher::new(
                                 builder.task_executor().clone(),
@@ -90,24 +95,14 @@ fn main() {
                         .await?;
 
                     handle.node_exit_future.await
-                }
+                },
                 true => {
                     let handle =
-                        builder
-                            .node(OpRbuilderNode::new(op_rbuilder_args.clone())).extend_rpc_modules(move |ctx| {
-                            // register eth bundle api
-                            let ext = EthBundleMinimalApi::new(ctx.registry.pool().clone());
-                            ctx.modules.merge_configured(ext.into_rpc())?;
+                        builder.node(OpNode::new(rollup_args.clone())).launch().await?;
 
-                            Ok(())
-                            })
-                            .launch().await?;
                     handle.node_exit_future.await
-                }
+                },
             }
         })
-        {
-            eprintln!("Error: {err:?}");
-            std::process::exit(1);
-        }
+        .unwrap();
 }
